feature(variables): extracted `createGroqBuilder` into separate file
