+++ packages/groq-builder/package.json
@@ -19,7 +19,8 @@
   ],
   "main": "./dist/index.js",
   "sideEffects": [
-    "./dist/commands/**"
+    "./dist/commands/**",
+    "./dist/groq-builder"
   ],
   "module": "dist/index.mjs",
   "types": "dist/index.d.ts",
