+++ packages/groq-builder/src/commands/conditional-types.ts
@@ -5,8 +5,8 @@ import {
 } from "./projection-types";
 import { Empty, IntersectionOfValues, Simplify, ValueOf } from "../types/utils";
 import { ExtractTypeNames, QueryConfig } from "../types/schema-types";
-import { GroqBuilder } from "../groq-builder";
-import { IGroqBuilder, InferResultType } from "../types/public-types";
+import { GroqBuilder, IGroqBuilder } from "../groq-builder";
+import { InferResultType } from "../types/public-types";
 import { Expressions } from "../types/groq-expressions";
 
 export type ConditionalProjectionMap<
