+++ packages/groq-builder/src/commands/conditional.test.ts
@@ -1,13 +1,9 @@
 import { describe, expect, expectTypeOf, it } from "vitest";
-import {
-  createGroqBuilder,
-  GroqBuilder,
-  InferResultItem,
-  InferResultType,
-} from "../index";
+import { GroqBuilder, InferResultItem, InferResultType } from "../index";
 import { SchemaConfig } from "../tests/schemas/nextjs-sanity-fe";
 import { ExtractConditionalProjectionTypes } from "./conditional-types";
 import { Empty, Simplify } from "../types/utils";
+import { createGroqBuilder } from "../createGroqBuilder";
 
 const q = createGroqBuilder<SchemaConfig>({ indent: "  " });
 const qVariants = q.star.filterByType("variant");
