+++ packages/groq-builder/src/commands/projection-types.ts
@@ -1,4 +1,4 @@
-import { GroqBuilder } from "../groq-builder";
+import { GroqBuilder, IGroqBuilder } from "../groq-builder";
 import {
   Empty,
   IsAny,
@@ -12,7 +12,6 @@ import {
 } from "../types/utils";
 import {
   FragmentInputTypeTag,
-  IGroqBuilder,
   Parser,
   ParserWithWidenedInput,
 } from "../types/public-types";
