+++ packages/groq-builder/src/commands/select-types.ts
@@ -1,7 +1,7 @@
 import { ExtractTypeNames, QueryConfig } from "../types/schema-types";
 import { StringKeys, ValueOf } from "../types/utils";
-import { GroqBuilder } from "../groq-builder";
-import { IGroqBuilder, InferResultType } from "../types/public-types";
+import { GroqBuilder, IGroqBuilder } from "../groq-builder";
+import { InferResultType } from "../types/public-types";
 import { Expressions } from "../types/groq-expressions";
 
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
