+++ packages/groq-builder/src/commands/select.ts
@@ -2,11 +2,7 @@ import { GroqBuilder } from "../groq-builder";
 import { ResultItem } from "../types/result-types";
 import { ExtractSelectResult, SelectProjections } from "./select-types";
 import { notNull } from "../types/utils";
-import {
-  IGroqBuilder,
-  InferResultType,
-  ParserFunction,
-} from "../types/public-types";
+import { InferResultType, ParserFunction } from "../types/public-types";
 
 declare module "../groq-builder" {
   export interface GroqBuilder<TResult, TQueryConfig> {
